+++ pytsite/contact_form/__init__.py
@@ -1,7 +1,7 @@
 """PytSite Contact Form.
 """
 # Public API
-from ._form import Contact as ContactForm
+from ._form import Form
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -12,10 +12,9 @@ def __init():
     from pytsite import assetman, lang, tpl, browser
 
     assetman.register_package(__name__)
-    assetman.add('pytsite.contact@js/common.js', forever=True)
     lang.register_package(__name__)
     tpl.register_package(__name__)
-    browser.register_ep('pytsite.contact.ep.submit')
+    browser.register_ep('pytsite.contact_form.ep.submit')
 
 
 __init()