+++ pytsite/contact_form/res/assets/js/common.js
@@ -2,11 +2,11 @@ $(function () {
     $('#pytsite-contact-form').on('pytsite_form_submit', function (e, form) {
         form.removeClass('ready-to-submit');
 
-        pytsite.browser.post('pytsite.contact.ep.submit', form.serializeForm(), function(response) {
+        pytsite.browser.post('pytsite.contact_form.ep.submit', form.serializeForm(), function(response) {
             alert(response);
             form[0].reset();
         }, function() {
-            alert(t('pytsite.contact@error_occurred'));
+            alert(t('pytsite.contact_form@error_occurred'));
         });
     });
 });